+++ CHANGELOG.md
@@ -31,7 +31,8 @@ We haven't had the best track record of code/feature changes before this date, b
      Unreleased changes that have landed in master. Click to see more.
     </summary>
 
--  Fix version in header commments
+- [refactor] Flow type strict substituteTokens.js
+- Fix version in header comments
 
   </details>